Crate rapier2d

Rapier

Rapier is a set of two Rust crates rapier2d and rapier3d for efficient cross-platform physics simulation. It target application include video games, animation, robotics, etc.

Rapier has some unique features for collaborative applications:

The ability to snapshot the state of the physics engine, and restore it later.

The ability to run a perfectly deterministic simulation on different machine, as long as they are compliant with the IEEE 754-2008 floating point standard.

Re-exports

pub extern crate parry2d as parry;

pub extern crate crossbeam;

pub extern crate nalgebra as na;

Modules

counters
Counters for benchmarking various parts of the physics engine.

data
Data structures modified with guaranteed deterministic behavior after deserialization.

dynamics
Structures related to dynamics: bodies, joints, etc.

geometry
Structures related to geometry: colliders, shapes, etc.

math
Elementary mathematical entities (vectors, matrices, isometries, etc).

pipeline
Structure for combining the various physics components to perform an actual simulation.

prelude
Prelude containing the common types defined by Rapier.

utils
Miscellaneous utilities.

Constants

VERSION
The string version of Rapier.