rapier2d
Module counters

Counters for benchmarking various parts of the physics engine.

Structs

CCDCounters

Performance counters related to continuous collision detection (CCD).

CollisionDetectionCounters

Performance counters related to collision detection.

Counters

Aggregation of all the performances counters tracked by rapier.

SolverCounters

Performance counters related to constraints resolution.

StagesCounters

Performance counters related to each stage of the time step.

Timer

A timer.