Module rapier2d::data

Data structures modified with guaranteed deterministic behavior after deserialization.

Re-exports

pub use self::arena::Arena;

pub use self::arena::Index;

Modules

arena
Arena adapted from the generational-arena crate.

pubsub
Publish-subscribe mechanism for internal events.

Structs

Coarena
A container for data associated to item existing into another Arena.

Traits

BundleSet
Helper trait to address multiple elements at once.

ComponentSet
A set of elements of type T.

ComponentSetMut
A set of mutable elements of type T.

ComponentSetOption
A set of optional elements of type T.