Module rapier2d::data::arena

Arena adapted from the generational-arena crate.

See https://github.com/fitzgen/generational-arena/blob/master/src/lib.rs. This has been modified to have a fully deterministic deserialization (including for the order of Index attribution after a deserialization of the arena).

Structs

Arena

The Arena allows inserting and removing elements that are referred to by Index.

Drain

An iterator that removes elements from the arena.

Index

An index (and generation) into an Arena.

IntoIter

An iterator over the elements in an arena.

Iter

An iterator over shared references to the elements in an arena.

IterMut

An iterator over exclusive references to elements in this arena.