Module rapier2d::dynamics

Structures related to dynamics: bodies, impulse_joints, etc.

Structs

BodyPair

A pair of rigid body handles.

CCDSolver

Solver responsible for performing motion-clamping on fast-moving bodies.

FixedJoint

A fixed joint, locks all relative motion between two bodies.

FixedJointBuilder

Create fixed joints using the builder pattern.

GenericJoint

A generic joint.

GenericJointBuilder

Create generic joints using the builder pattern.

ImpulseJoint

An impulse-based joint attached to two bodies.

ImpulseJointHandle

The unique identifier of a joint added to the joint set. The unique identifier of a collider added to a collider set.

ImpulseJointSet

A set of impulse_joints that can be handled by a physics World.

IntegrationParameters

Parameters for a time-step of the physics engine.

IslandManager

Structure responsible for maintaining the set of active rigid-bodies, and putting non-moving rigid-bodies to sleep to save computation times.

JointAxesMask

A bit mask identifying multiple degrees of freedom of a joint.

JointLimits

The limits of a joint along one of its degrees of freedom.

JointMotor

A joint’s motor along one of its degrees of freedom.

LockedAxes

Flags affecting the behavior of the constraints solver for a given contact manifold.

MassProperties

The local mass properties of a rigid-body.

Multibody

An articulated body simulated using the reduced-coordinates approach.

MultibodyIndex

The temporary index of a multibody added to a MultibodyJointSet.

MultibodyJoint

An joint attached to two bodies based on the reduced coordinates formalism.

MultibodyJointHandle

The unique handle of an multibody_joint added to a MultibodyJointSet.

MultibodyJointSet

A set of rigid bodies that can be handled by a physics pipeline.

MultibodyLink

One link of a multibody.

MultibodyLinkId

Indexes usable to get a multibody link from a MultibodyJointSet.

PrismaticJoint

A prismatic joint, locks all relative motion between two bodies except for translation along the joint’s principal axis.

PrismaticJointBuilder

Create prismatic joints using the builder pattern.

RevoluteJoint

A revolute joint, locks all relative motion except for rotation along the joint’s principal axis.

RevoluteJointBuilder

Create revolute joints using the builder pattern.

RigidBody

A rigid body.

RigidBodyActivation

The rb_activation status of a body.

RigidBodyBuilder

A builder for rigid-bodies.

RigidBodyCcd

Information used for Continuous-Collision-Detection.

RigidBodyChanges

Flags describing how the rigid-body has been modified by the user.

RigidBodyColliders

The set of colliders attached to this rigid-bodies.

RigidBodyDamping

Damping factors to progressively slow down a rigid-body.

RigidBodyDominance

The dominance groups of a rigid-body.

RigidBodyForces

The user-defined external forces applied to this rigid-body.

RigidBodyHandle

The unique handle of a rigid body added to a RigidBodySet.

RigidBodyIds

Internal identifiers used by the physics engine.

RigidBodyMassProps

The mass properties of a rigid-body.

RigidBodyPosition

The position of this rigid-body.

RigidBodySet

A set of rigid bodies that can be handled by a physics pipeline.

RigidBodyVelocity

The velocities of this rigid-body.

RopeJoint

A rope joint, limits the maximum distance between two bodies

RopeJointBuilder

Create rope joints using the builder pattern.

SpringJoint

A spring-damper joint, applies a force proportional to the distance between two objects.

SpringJointBuilder

A SpringJoint joint using the builder pattern.

Enums

CoefficientCombineRule

Rules used to combine two coefficients.

JointAxis

Identifiers of degrees of freedoms of a joint.

JointEnabled

Enum indicating whether or not a joint is enabled.

MotorModel

The spring-like model used for constraints resolution.

RigidBodyAdditionalMassProps

Mass and angular inertia added to a rigid-body on top of its attached colliders’ contributions.

RigidBodyType

The status of a body, governing the way it is affected by external forces.

Functions

unit_joint_limit_constraint

Initializes and generate the velocity constraints applicable to the multibody links attached to this multibody_joint.

unit_joint_motor_constraint

Initializes and generate the velocity constraints applicable to the multibody links attached to this multibody_joint.

Type Aliases

BodyStatusDeprecated

The type of a body, governing the way it is affected by external forces.