Struct rapier2d::dynamics::IntegrationParameters

pub struct IntegrationParameters {
    pub dt: Real,
    pub min_ccd_dt: Real,
    pub erp: Real,
    pub joint_erp: Real,
    pub warmstart_coeff: Real,
    pub warmstart_correction_slope: Real,
    pub velocity_solve_fraction: Real,
    pub velocity_based_erp: Real,
    pub allowed_linear_error: Real,
    pub prediction_distance: Real,
    pub allowed_angular_error: Real,
    pub max_linear_correction: Real,
    pub max_angular_correction: Real,
    pub max_velocity_iterations: usize,
    pub max_position_iterations: usize,
    pub min_island_size: usize,
    pub max_ccd_substeps: usize,
}

Parameters for a time-step of the physics engine.

Fields
dt: Real

The timestep length (default: 1.0 / 60.0)

min_ccd_dt: Real

Minimum timestep size when using CCD with multiple substeps (default 1.0 / 60.0 / 100.0)

When CCD with multiple substeps is enabled, the timestep is subdivided into smaller pieces. This timestep subdivision won't generate timestep lengths smaller than min_ccd_dt.

Setting this to a large value will reduce the opportunity to performing CCD substepping, resulting in potentially more time dropped by the motion-clamping mechanism. Setting this to an very small value may lead to numerical instabilities.

erp: Real

The Error Reduction Parameter in [0, 1] is the proportion of the positional error to be corrected at each time step (default: 0.2).

joint_erp: Real

The Error Reduction Parameter for joints in [0, 1] is the proportion of the positional error to be corrected at each time step (default: 0.2).

warmstart_coeff: Real

Each cached impulse are multiplied by this coefficient in [0, 1] when they are re-used to initialize the solver (default 1.0).

warmstart_correction_slope: Real

Correction factor to avoid large warmstart impulse after a strong impact (default 10.0).

velocity_solve_fraction: Real

0-1: how much of the velocity to dampen out in the constraint solver? (default 1.0).

velocity_based_erp: Real

0-1: multiplier for how much of the constraint violation (e.g. contact penetration) will be compensated for during the velocity solve. If zero, you need to enable the positional solver. If non-zero, you do not need the positional solver. A good non-zero value is around 0.2. (default 0.0).

allowed_linear_error: Real

Amount of penetration the engine wont attempt to correct (default: 0.005m).

prediction_distance: Real

The maximal distance separating two objects that will generate predictive contacts (default: 0.002).

allowed_angular_error: Real

Amount of angular drift of joint limits the engine wont attempt to correct (default: 0.001rad).

max_linear_correction: Real

Maximum linear correction during one step of the non-linear position solver (default: 0.2).

max_angular_correction: Real

Maximum angular correction during one step of the non-linear position solver (default: 0.2).

max_velocity_iterations: usize

Maximum number of iterations performed by the velocity constraints solver (default: 4).

max_position_iterations: usize

Maximum number of iterations performed by the position-based constraints solver (default: 1).

min_island_size: usize

Minimum number of dynamic bodies in each active island (default: 128).

max_ccd_substeps: usize

Maximum number of substeps performed by the solver (default: 1).

Implementations

impl IntegrationParameters

pub fn new(
dt: Real,
erp: Real,
joint_erp: Real,
warmstart_coeff: Real,
allowed_linear_error: Real,
allowed_angular_error: Real,
max_linear_correction: Real,
max_angular_correction: Real,
prediction_distance: Real,
max_velocity_iterations: usize,
max_position_iterations: usize,
max_ccd_substeps: usize
) -> Self

👎 Deprecated:
Use IntegrationParameters { dt: 60.0, ..Default::default() } instead

Creates a set of integration parameters with the given values.

pub fn dt(&self) -> Real

👎 Deprecated:
You can just read the IntegrationParams::dt value directly

The current time-stepping length.

pub fn inv_dt(&self) -> Real

The inverse of the time-stepping length, i.e. the steps per seconds (Hz).

This is zero if self.dt is zero.

pub fn set_dt(&mut self, dt: Real)

👎 Deprecated:
You can just set the IntegrationParams::dt value directly

Sets the time-stepping length.

pub fn set_inv_dt(&mut self, inv_dt: Real)

Sets the inverse time-stepping length (i.e. the frequency).

This automatically recompute self.dt.