Struct rapier2d::dynamics::JointSet

pub struct JointSet { /* fields omitted */ }

A set of joints that can be handled by a physics World.

Implementations

impl JointSet

pub fn new() -> Self

Creates a new empty set of joints.

pub fn len(&self) -> usize

The number of joints on this set.

pub fn is_empty(&self) -> bool

true if there are no joints in this set.

pub fn joint_graph(&self) -> &InteractionGraph<RigidBodyHandle, Joint>

Retrieve the joint graph where edges are joints and nodes are rigid body handles.

pub fn joints_with<'a>(
&'a self,
body: RigidBodyHandle
) -> impl Iterator<Item = (RigidBodyHandle, RigidBodyHandle, &'a Joint)>

Iterates through all the joitns attached to the given rigid-body.

pub fn contains(&self, handle: JointHandle) -> bool

Is the given joint handle valid?

pub fn get(&self, handle: JointHandle) -> Option<&Joint>

Gets the joint with the given handle.

pub fn get_mut(&mut self, handle: JointHandle) -> Option<&mut Joint>

Gets a mutable reference to the joint with the given handle.

pub fn get_unknown_gen(&self, i: u32) -> Option<(&Joint, JointHandle)>

Gets the joint with the given handle without a known generation.

This is useful when you know you want the joint at position i but don't know what is its current generation number. Generation numbers are used to protect from the ABA problem because the joint position i are recycled between two insertion and a removal.

Using this is discouraged in favor of self.get(handle) which does not suffer form the ABA problem.

pub fn get_unknown_gen_mut(
&mut self,
i: u32
) -> Option<(&mut Joint, JointHandle)>

Gets a mutable reference to the joint with the given handle without a known generation.

This is useful when you know you want the joint at position i but don't know what is its current generation number. Generation numbers are used to protect from the ABA problem because the joint position i are recycled between two insertion and a removal.

Using this is discouraged in favor of self.get_mut(handle) which does not suffer form the ABA problem.

pub fn iter(&self) -> impl Iterator<Item = (JointHandle, &Joint)>

Iterates through all the joint on this set.

pub fn iter_mut(&mut self) -> impl Iterator<Item = (JointHandle, &mut Joint)>

Iterates mutably through all the joint on this set.

pub fn insert<J>(
&mut self,
_bodies: &mut impl ComponentSetMut<RigidBodyIds>,
body1: RigidBodyHandle,
body2: RigidBodyHandle,
joint_params: J
) -> JointHandle where
J: Into<JointParams>,

Inserts a new joint into this set and retrieve its handle.

pub fn remove<Bodies>(
&mut self,
handle: JointHandle,
islands: &mut IslandManager,
bodies: &mut Bodies,
wake_up: bool
) -> Option<Joint> where
Bodies: ComponentSetMut<RigidBodyActivation> + ComponentSet<RigidBodyType> + ComponentSetMut<RigidBodyIds>,

Removes a joint from this set.

If wake_up is set to true, then the bodies attached to this joint will be automatically woken up.

pub fn remove_joints_attached_to_rigid_body<Bodies>(
&mut self,
handle: RigidBodyHandle,
islands: &mut IslandManager,
bodies: &mut Bodies
) -> Vec<JointHandle> where
Bodies: ComponentSetMut<RigidBodyActivation> + ComponentSet<RigidBodyType> + ComponentSetMut<RigidBodyIds>,

Deletes all the joints attached to the given rigid-body.

The provided rigid-body handle is not required to identify a rigid-body that is still contained by the bodies component set. Returns the (now invalid) handles of the removed joints.