Struct rapier2d::dynamics::MassProperties

pub struct MassProperties {
    pub local_com: Point<f32, 2_usize>,
    pub inv_mass: f32,
    pub inv_principal_inertia_sqrt: f32,
}

The local mass properties of a rigid-body.

Fields
local_com: Point<f32, 2_usize>

The center of mass of a rigid-body expressed in its local-space.

inv_mass: f32

The inverse of the mass of a rigid-body.

If this is zero, the rigid-body is assumed to have infinite mass.

inv_principal_inertia_sqrt: f32

The inverse of the principal angular inertia of the rigid-body.

Components set to zero are assumed to be infinite along the corresponding principal axis.

Implementations

impl MassProperties

pub fn new(
local_com: Point<f32, 2_usize>,
mass: f32,
principal_inertia: f32
) -> MassProperties

Initializes the mass properties with the given center-of-mass, mass, and angular inertia.

The center-of-mass is specified in the local-space of the rigid-body.

pub fn world_com(
&self,
pos: &Isometry<f32, Unit<Complex<f32>>, 2_usize>
) -> Point<f32, 2_usize>

The world-space center of mass of the rigid-body.

pub fn world_inv_inertia_sqrt(&self, _rot: &Unit<Complex<f32>>) -> f32

The world-space inverse angular inertia tensor of the rigid-body.

pub fn transform_by(
&self,
m: &Isometry<f32, Unit<Complex<f32>>, 2_usize>
) -> MassProperties

Transform each element of the mass properties.

pub fn set_mass(&mut self, new_mass: f32, adjust_angular_inertia: bool)

Changes the mass on these mass-properties.

The adjust_angular_inertia argument should always be true, unless there are some specific reasons not to do so. Setting this to true will automatically adjust the angular inertia of self to account for the mass change (i.e. it will multiply the angular inertia by new_mass / prev_mass). Setting it to false will not change the current angular inertia.

impl MassProperties

pub fn from_ball(density: f32, radius: f32) -> MassProperties

Computes the mass properties of a ball.

impl MassProperties

pub fn from_capsule(
density: f32,
a: Point<f32, 2_usize>,
b: Point<f32, 2_usize>,
radius: f32
) -> MassProperties

Computes the mass properties of a capsule.

impl MassProperties

pub fn from_compound(
density: f32,
shapes: &[(Isometry<f32, Unit<Complex<f32>>, 2_usize>, SharedShape)]
) -> MassProperties

Computes the mass properties of a compound shape.

impl MassProperties

pub fn from_convex_polygon(
density: f32,
vertices: &[Point<f32, 2_usize>]
) -> MassProperties

Computes the mass properties of a convex polygon.

impl MassProperties

pub fn from_cuboid(
density: f32,
half_extents: Matrix<f32, Const<{_: usize}>, Const<1_usize>, ArrayStorage<f32, 2_usize, 1_usize>>
) -> MassProperties

Computes the mass properties of a cuboid.

impl MassProperties

pub fn from_triangle(
density: f32,
a: &Point<f32, 2_usize>,
b: &Point<f32, 2_usize>,
c: &Point<f32, 2_usize>
) -> MassProperties

Computes the mass properties of a triangle.

impl MassProperties

pub fn from_trimesh(
density: f32,
vertices: &[Point<f32, 2_usize>],
indices: &[[u32; 3]]
) -> MassProperties

Computes the mass properties of a triangle-mesh.