Module rapier2d::geometry

Structures related to geometry: colliders, shapes, etc.

Structs

ActiveCollisionTypes
Flags affecting whether or not collision-detection happens between two colliders depending on the type of rigid-bodies they are attached to.

Ball
A Ball shape.

BroadPhase
A broad-phase combining a Hierarchical Grid and Sweep-and-Prune.

Capsule
A capsule shape defined as a round segment.

ColliderBroadPhaseData
Data associated to a collider that takes part to a broad-phase algorithm.

ColliderChanges
Flags describing how the collider has been modified by the user.

ColliderFlags
A set of flags for controlling collision/intersection filtering, modification, and events.

ColliderHandle
The unique identifier of a collider added to a collider set.

ColliderMaterial
The constraints solver-related properties of this collider (friction, restitution, etc.)

ColliderParent
Information about the rigid-body this collider is attached to.

ColliderPosition
The position of a collider.

Compound
A compound shape with an aabb bounding volume.

ContactData
A single contact between two collider.

ContactManifoldData
A contact manifold between two colliders.

ContactPair
The description of all the contacts between a pair of colliders.

ConvexPolygon
A 2D convex polygon.

Cuboid
Shape of a box.

HalfSpace
A half-space delimited by an infinite plane.

HeightField
A 2D heightfield.

InteractionGraph
A graph where nodes are collision objects and edges are contact or proximity algorithms.

InteractionGroups
Pairwise filtering using bit masks.

IntersectionEvent
Events occurring when two collision objects start or stop being in close proximity, contact, or disjoint.

NarrowPhase
The narrow-phase responsible for computing precise contact information between colliders.

PolygonalFeature
A polygonal feature representing the local polygonal approximation of a vertex, or face, of a convex shape.

Polyline
A polyline.

RoundShape
A shape with rounded borders.

Segment
A segment shape.

SharedShape
The shape of a collider.

SolverContact
A contact seen by the constraints solver for computing forces.

SolverFlags
Flags affecting the behavior of the constraints solver for a given contact manifold.

TrackedContact
A single contact between two shape.

TriMesh
A triangle mesh.

Triangle
A triangle shape.

Enums

ColliderMassProps
The mass-properties of a collider.

ColliderType
The type of collider.

ContactEvent
Events occurring when two collision objects start or stop being in contact (or penetration).

FeatureId
An identifier of a feature of a convex polyhedron.

SegmentPointLocation
Logical description of the location of a point on a triangle.

ShapeType
Enum representing the type of a shape.

TrianglePointLocation
Description of the location of a point on a triangle.

TypedShape
Enum representing the shape with its actual type

Traits

PolygonalFeatureMap
Trait implemented by convex shapes with features with polyhedral approximations.

Shape
Trait implemented by shapes usable by Rapier.

SimdCompositeShape
Trait implemented by shapes composed of multiple simpler shapes.

SupportMap
Traits of convex shapes representable by a support mapping function.

TypedSimdCompositeShape

Type Definitions

AABB
An axis-aligned bounding box.

Ball
A ball shape.

Capsule
A capsule shape.

ColliderGraphIndex
Index of a node of the interaction graph.

ColliderShape
The shape of a collider.

Contact
A contact between two colliders.

ContactManifold
A contact manifold between two colliders.

Cuboid
A cuboid shape.

HeightField
A heightfield shape.

PointProjection
The the projection of a point on a collider.

Ray
A ray that can be cast against colliders.

RayIntersection
The intersection between a ray and a collider.

RigidBodyGraphIndex
Index of a node of the interaction graph.

RoundConvexPolygon
A convex polygon dilated by a sphere (so it has round corners).

RoundCuboid
A cuboid dilated by a sphere (so it has round corners).

RoundTriangle
A triangle dilated by a sphere (so it has round corners).

Segment
A segment shape.

TOI
The the time of impact between two shapes.

TemporaryInteractionIndex
Temporary index to and edge of the interaction graph.

Triangle
A triangle shape.