rapier2d::geometry
Struct NarrowPhase

pub struct NarrowPhase { /* private fields */ }

The narrow-phase responsible for computing precise contact information between colliders.

Implementations

impl NarrowPhase

pub fn new() -> Self

Creates a new empty narrow-phase.

pub fn with_query_dispatcher<D>(d: D) -> Self
where D: 'static + PersistentQueryDispatcher<ContactManifoldData, ContactData>,

Creates a new empty narrow-phase with a custom query dispatcher.

pub fn query_dispatcher( &self, ) -> &dyn PersistentQueryDispatcher<ContactManifoldData, ContactData>

The query dispatcher used by this narrow-phase to select the right collision-detection algorithms depending of the shape types.

pub fn contact_graph(&self) -> &InteractionGraph<ColliderHandle, ContactPair>

The contact graph containing all contact pairs and their contact information.

pub fn intersection_graph( &self, ) -> &InteractionGraph<ColliderHandle, IntersectionPair>

The intersection graph containing all intersection pairs and their intersection information.

pub fn contact_pairs_with_unknown_gen( &self, collider: u32, ) -> impl Iterator<Item = &ContactPair>

All the contacts involving the given collider.

It is strongly recommended to use the NarrowPhase::contact_pairs_with method instead. This method can be used if the generation number of the collider handle isnโt known.

pub fn contact_pairs_with( &self, collider: ColliderHandle, ) -> impl Iterator<Item = &ContactPair>

All the contact pairs involving the given collider.

The returned contact pairs identify pairs of colliders with intersecting bounding-volumes. To check if any geometric contact happened between the collider shapes, check ContactPair::has_any_active_contact.

pub fn intersection_pairs_with_unknown_gen( &self, collider: u32, ) -> impl Iterator<Item = (ColliderHandle, ColliderHandle, bool)> + '_

All the intersection pairs involving the given collider.

It is strongly recommended to use the NarrowPhase::intersection_pairs_with method instead. This method can be used if the generation number of the collider handle isnโt known.

pub fn intersection_pairs_with( &self, collider: ColliderHandle, ) -> impl Iterator<Item = (ColliderHandle, ColliderHandle, bool)> + '_

All the intersection pairs involving the given collider, where at least one collider involved in the intersection is a sensor.

The returned contact pairs identify pairs of colliders (where at least one is a sensor) with intersecting bounding-volumes. To check if any geometric overlap happened between the collider shapes, check the returned boolean.

pub fn contact_pair_at_index( &self, id: TemporaryInteractionIndex, ) -> &ContactPair

Returns the contact pair at the given temporary index.

pub fn contact_pair_unknown_gen( &self, collider1: u32, collider2: u32, ) -> Option<&ContactPair>

The contact pair involving two specific colliders.

It is strongly recommended to use the NarrowPhase::contact_pair method instead. This method can be used if the generation number of the collider handle isnโt known.

If this returns None, there is no contact between the two colliders. If this returns Some, then there may be a contact between the two colliders. Check the result ContactPair::has_any_active_contact method to see if there is an actual contact.

pub fn contact_pair( &self, collider1: ColliderHandle, collider2: ColliderHandle, ) -> Option<&ContactPair>

The contact pair involving two specific colliders.

If this returns None, there is no contact between the two colliders. If this returns Some, then there may be a contact between the two colliders. Check the result ContactPair::has_any_active_contact method to see if there is an actual contact.

pub fn intersection_pair_unknown_gen( &self, collider1: u32, collider2: u32, ) -> Option<bool>

The intersection pair involving two specific colliders.

It is strongly recommended to use the NarrowPhase::intersection_pair method instead. This method can be used if the generation number of the collider handle isnโt known.

If this returns None or Some(false), then there is no intersection between the two colliders. If this returns Some(true), then there may be an intersection between the two colliders.

pub fn intersection_pair( &self, collider1: ColliderHandle, collider2: ColliderHandle, ) -> Option<bool>

The intersection pair involving two specific colliders.

If this returns None or Some(false), then there is no intersection between the two colliders. If this returns Some(true), then there may be an intersection between the two colliders.

pub fn contact_pairs(&self) -> impl Iterator<Item = &ContactPair>

All the contact pairs maintained by this narrow-phase.

pub fn intersection_pairs( &self, ) -> impl Iterator<Item = (ColliderHandle, ColliderHandle, bool)> + '_

All the intersection pairs maintained by this narrow-phase.

pub fn handle_user_changes( &mut self, islands: Option<&mut IslandManager>, modified_colliders: &[ColliderHandle], removed_colliders: &[ColliderHandle], colliders: &mut ColliderSet, bodies: &mut RigidBodySet, events: &dyn EventHandler, )

Maintain the narrow-phase internal state by taking collider removal into account.