Trait rapier2d::geometry::Shape

pub trait Shape: PointQuery + RayCast + DowncastSync {
    pub fn compute_local_aabb(&self) -> AABB;

    pub fn compute_local_bounding_sphere(&self) -> BoundingSphere;

    pub fn clone_box(&self) -> Box<dyn Shape + 'static, Global>;

    pub fn mass_properties(&self, density: f32) -> MassProperties;

    pub fn shape_type(&self) -> ShapeType;

    pub fn as_typed_shape(&self) -> TypedShape<'_>;

    pub fn ccd_thickness(&self) -> f32;

    pub fn ccd_angular_thickness(&self) -> f32;

    pub fn compute_aabb(
        &self, 
        position: &Isometry<f32, Unit<Complex<f32>>, 2_usize>
    ) -> AABB { ... }

    pub fn compute_bounding_sphere(
        &self, 
        position: &Isometry<f32, Unit<Complex<f32>>, 2_usize>
    ) -> BoundingSphere { ... }

    pub fn is_convex(&self) -> bool { ... }

    pub fn as_support_map(&self) -> Option<&dyn SupportMap> { ... }

    pub fn as_composite_shape(&self) -> Option<&dyn SimdCompositeShape> { ... }

    pub fn as_polygonal_feature_map(
        &self
    ) -> Option<(&dyn PolygonalFeatureMap, f32)> { ... }

    pub fn feature_normal_at_point(
        &self, 
        _feature: FeatureId, 
        _point: &Point<f32, 2_usize>
    ) -> Option<Unit<Matrix<f32, Const<{_: usize}>, Const<1_usize>, ArrayStorage<f32, 2_usize, 1_usize>>>> { ... }

    pub fn compute_swept_aabb(
        &self, 
        start_pos: &Isometry<f32, Unit<Complex<f32>>, 2_usize>, 
        end_pos: &Isometry<f32, Unit<Complex<f32>>, 2_usize>
    ) -> AABB { ... }
}

Trait implemented by shapes usable by Rapier.

Required methods

pub fn compute_local_aabb(&self) -> AABB

Computes the AABB of this shape.

pub fn compute_local_bounding_sphere(&self) -> BoundingSphere

Computes the bounding-sphere of this shape.

pub fn clone_box(&self) -> Box<dyn Shape + 'static, Global>

Clones this shape into a boxed trait-object.

pub fn mass_properties(&self, density: f32) -> MassProperties

Compute the mass-properties of this shape given its uniform density.

pub fn shape_type(&self) -> ShapeType

Gets the type tag of this shape.

pub fn as_typed_shape(&self) -> TypedShape<'_>

Gets the underlying shape as an enum.

pub fn ccd_thickness(&self) -> f32

pub fn ccd_angular_thickness(&self) -> f32

Loading content...
Provided methods

pub fn compute_aabb(
&self,
position: &Isometry<f32, Unit<Complex<f32>>, 2_usize>
) -> AABB

Computes the AABB of this shape with the given position.

pub fn compute_bounding_sphere(
&self,
position: &Isometry<f32, Unit<Complex<f32>>, 2_usize>
) -> BoundingSphere

Computes the bounding-sphere of this shape with the given position.

pub fn is_convex(&self) -> bool

Is this shape known to be convex?

If this returns true then self is known to be convex. If this returns false then it is not known whether or not self is convex.

pub fn as_support_map(&self) -> Option<&dyn SupportMap>

Convents this shape into its support mapping, if it has one.

pub fn as_composite_shape(&self) -> Option<&dyn SimdCompositeShape>

pub fn as_polygonal_feature_map(
&self
) -> Option<(&dyn PolygonalFeatureMap, f32)>

Converts this shape to a polygonal feature-map, if it is one.

pub fn feature_normal_at_point(
&self,
_feature: FeatureId,
_point: &Point<f32, 2_usize>
) -> Option<Unit<Matrix<f32, Const<{_: usize}>, Const<1_usize>, ArrayStorage<f32, 2_usize, 1_usize>>>>

The shape's normal at the given point located on a specific feature.

pub fn compute_swept_aabb(
&self,
start_pos: &Isometry<f32, Unit<Complex<f32>>, 2_usize>,
end_pos: &Isometry<f32, Unit<Complex<f32>>, 2_usize>
) -> AABB

Computes the swept AABB of this shape, i.e., the space it would occupy by moving from the given start position to the given end position.

Loading content...
Implementations

impl dyn Shape + 'static

pub fn as_shape<T>(&self) -> Option<&T> where
T: Shape,

Converts this abstract shape to the given shape, if it is one.

pub fn as_shape_mut<T>(&mut self) -> Option<&mut T> where
T: Shape,

Converts this abstract shape to the given mutable shape, if it is one.

pub fn as_ball(&self) -> Option<&Ball>

Converts this abstract shape to a ball, if it is one.

pub fn as_ball_mut(&mut self) -> Option<&mut Ball>

Converts this abstract shape to a mutable ball, if it is one.

pub fn as_cuboid(&self) -> Option<&Cuboid>

Converts this abstract shape to a cuboid, if it is one.

pub fn as_cuboid_mut(&mut self) -> Option<&mut Cuboid>

Converts this abstract shape to a mutable cuboid, if it is one.

pub fn as_halfspace(&self) -> Option<&HalfSpace>

Converts this abstract shape to a halfspace, if it is one.

pub fn as_halfspace_mut(&mut self) -> Option<&mut HalfSpace>

Converts this abstract shape to a halfspace, if it is one.

pub fn as_segment(&self) -> Option<&Segment>

Converts this abstract shape to a segment, if it is one.

pub fn as_segment_mut(&mut self) -> Option<&mut Segment>

Converts this abstract shape to a mutable segment, if it is one.

pub fn as_capsule(&self) -> Option<&Capsule>

Converts this abstract shape to a capsule, if it is one.

pub fn as_capsule_mut(&mut self) -> Option<&mut Capsule>

Converts this abstract shape to a mutable capsule, if it is one.

pub fn as_triangle(&self) -> Option<&Triangle>

Converts this abstract shape to a triangle, if it is one.

pub fn as_triangle_mut(&mut self) -> Option<&mut Triangle>

Converts this abstract shape to a mutable triangle, if it is one.

pub fn as_compound(&self) -> Option<&Compound>

Converts this abstract shape to a compound shape, if it is one.

pub fn as_compound_mut(&mut self) -> Option<&mut Compound>

Converts this abstract shape to a mutable compound shape, if it is one.

pub fn as_trimesh(&self) -> Option<&TriMesh>

Converts this abstract shape to a triangle mesh, if it is one.

pub fn as_trimesh_mut(&mut self) -> Option<&mut TriMesh>

Converts this abstract shape to a mutable triangle mesh, if it is one.

pub fn as_polyline(&self) -> Option<&Polyline>

Converts this abstract shape to a polyline, if it is one.

pub fn as_polyline_mut(&mut self) -> Option<&mut Polyline>

Converts this abstract shape to a mutable polyline, if it is one.

pub fn as_heightfield(&self) -> Option<&HeightField>

Converts this abstract shape to a heightfield, if it is one.

pub fn as_heightfield_mut(&mut self) -> Option<&mut HeightField>

Converts this abstract shape to a mutable heightfield, if it is one.

pub fn as_round_cuboid(&self) -> Option<&RoundShape<Cuboid>>

Converts this abstract shape to a round cuboid, if it is one.

pub fn as_round_cuboid_mut(&mut self) -> Option<&mut RoundShape<Cuboid>>

Converts this abstract shape to a mutable round cuboid, if it is one.

pub fn as_round_triangle(&self) -> Option<&RoundShape<Triangle>>

Converts this abstract shape to a round triangle, if it is one.

pub fn as_round_triangle_mut(&mut self) -> Option<&mut RoundShape<Triangle>>

Converts this abstract shape to a round triangle, if it is one.

pub fn as_convex_polygon(&self) -> Option<&ConvexPolygon>

Converts this abstract shape to a convex polygon, if it is one.

pub fn as_convex_polygon_mut(&mut self) -> Option<&mut ConvexPolygon>

Converts this abstract shape to a mutable convex polygon, if it is one.

pub fn as_round_convex_polygon(&self) -> Option<&RoundShape<ConvexPolygon>>

Converts this abstract shape to a round convex polygon, if it is one.

pub fn as_round_convex_polygon_mut(
&mut self
) -> Option<&mut RoundShape<ConvexPolygon>>

Converts this abstract shape to a mutable round convex polygon, if it is one.

impl dyn Shape + 'static

pub fn is<__T>(&self) -> bool where
__T: Shape,

Returns true if the trait object wraps an object of type __T.

pub fn downcast<__T>(
self: Box<dyn Shape + 'static, Global>
) -> Result<Box<__T, Global>, Box<dyn Shape + 'static, Global>> where
__T: Shape,

Returns a boxed object from a boxed trait object if the underlying object is of type __T. Returns the original boxed trait if it isn't.

pub fn downcast_rc<__T>(
self: Rc<dyn Shape + 'static>
) -> Result<Rc<__T>, Rc<dyn Shape + 'static>> where
__T: Shape,

Returns an Rc-ed object from an Rc-ed trait object if the underlying object is of type __T. Returns the original Rc-ed trait if it isn't.

pub fn downcast_ref<__T>(&self) -> Option<&__T> where
__T: Shape,

Returns a reference to the object within the trait object if it is of type __T, or None if it isn't.

pub fn downcast_mut<__T>(&mut self) -> Option<&mut __T> where
__T: Shape,

Returns a mutable reference to the object within the trait object if it is of type __T, or None if it isn't.

pub fn downcast_arc<__T>(
self: Arc<dyn Shape + 'static>
) -> Result<Arc<__T>, Arc<dyn Shape + 'static>> where
__T: Shape + Any + Send + Sync,

Returns an Arc-ed object from an Arc-ed trait object if the underlying object is of type __T. Returns the original Arc-ed trait if it isn't.