Module rapier2d::pipeline

Structure for combining the various physics components to perform an actual simulation.

Structs

ActiveEvents

Flags affecting the events generated for this collider.

ActiveHooks

Flags affecting the behavior of the constraints solver for a given contact manifold.

ChannelEventCollector

A collision event handler that collects events into a crossbeam channel.

CollisionPipeline

The collision pipeline, responsible for performing collision detection between colliders.

ContactModificationContext

Context given to custom contact modifiers to modify the contacts seen by the constraints solver.

PairFilterContext

Context given to custom collision filters to filter-out collisions.

PhysicsPipeline

The physics pipeline, responsible for stepping the whole physics simulation.

QueryFilter

A filter tha describes what collider should be included or excluded from a scene query.

QueryFilterFlags

Flags for excluding whole sets of colliders from a scene query.

QueryPipeline

A pipeline for performing queries on all the colliders of a scene.

Enums

QueryPipelineMode

Indicates how the colliders position should be taken into account when updating the query pipeline.

Traits

EventHandler

Trait implemented by structures responsible for handling events generated by the physics engine.

PhysicsHooks

User-defined functions called by the physics engines during one timestep in order to customize its behavior.