Trait rapier2d::pipeline::EventHandler

pub trait EventHandler: Send + Sync {
    fn handle_intersection_event(&self, event: IntersectionEvent);

    fn handle_contact_event(
        &self, 
        event: ContactEvent, 
        contact_pair: &ContactPair
    );
}

Trait implemented by structures responsible for handling events generated by the physics engine.

Implementors of this trait will typically collect these events for future processing.

Required methods

fn handle_intersection_event(&self, event: IntersectionEvent)

Handle an intersection event.

A intersection event is emitted when the state of intersection between two colliders changes.

fn handle_contact_event(&self, event: ContactEvent, contact_pair: &ContactPair)

Handle a contact event.

A contact event is emitted when two collider start or stop touching, independently from the number of contact points involved.

Loading content...