rapier2d
Module utils

Miscellaneous utilities.

Traits

IndexMut2

Methods for simultaneously indexing a container with two distinct indices.

SimdBasis

Trait to compute the orthonormal basis of a vector.

SimdQuat

Trait implemented by quaternions.

SimdRealCopy

The trait for real numbers used by Rapier.

SimdSign

Trait to copy the sign of each component of one scalar/vector/matrix to another.

Functions

smallest_abs_diff_between_angles

Calculate the difference with smallest absolute value between the two given angles.

smallest_abs_diff_between_sin_angles

Calculate the difference with smallest absolute value between the two given values.