Trait rapier2d::utils::IndexMut2

pub trait IndexMut2<I>: IndexMut<I> {
    fn index_mut2(
        &mut self, 
        i: usize, 
        j: usize
    ) -> (&mut Self::Output, &mut Self::Output);

    fn index_mut_const(
        &mut self, 
        i: usize, 
        j: usize
    ) -> (&mut Self::Output, &Self::Output) { ... }
}

Methods for simultaneously indexing a container with two distinct indices.

Required methods

fn index_mut2(
&mut self,
i: usize,
j: usize
) -> (&mut Self::Output, &mut Self::Output)

Gets mutable references to two distinct elements of the container.

Panics if i == j.

Loading content...
Provided methods

fn index_mut_const(
&mut self,
i: usize,
j: usize
) -> (&mut Self::Output, &Self::Output)

Gets a mutable reference to one element, and immutable reference to a second one.

Panics if i == j.

Loading content...