rapier2d::utils
Trait SimdBasis

pub trait SimdBasis: Sized {
    type Basis;

    // Required methods
    fn orthonormal_basis(self) -> Self::Basis;
    fn orthonormal_vector(self) -> Self;
}

Trait to compute the orthonormal basis of a vector.

Required Associated Types

type Basis

The type of the array of orthonormal vectors.

Required Methods

fn orthonormal_basis(self) -> Self::Basis

Computes the vectors which, when combined with self, form an orthonormal basis.

fn orthonormal_vector(self) -> Self

Computes a vector orthogonal to self with a unit length (if self has a unit length).

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.