rapier2d::utils
Trait SimdSign

pub trait SimdSign<Rhs>: Sized {
    // Required method
    fn copy_sign_to(self, to: Rhs) -> Rhs;
}

Trait to copy the sign of each component of one scalar/vector/matrix to another.

Required Methods

fn copy_sign_to(self, to: Rhs) -> Rhs

Copy the sign of each component of self to the corresponding component of to.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.