Trait rapier2d::utils::WSign

pub trait WSign<Rhs>: Sized {
    fn copy_sign_to(self, to: Rhs) -> Rhs;
}

Trait to copy the sign of each component of one scalar/vector/matrix to another.

Required methods

fn copy_sign_to(self, to: Rhs) -> Rhs

Copy the sign of each component of self to the corresponding component of to.

Loading content...