Module rapier2d::dynamics

Structures related to dynamics: bodies, joints, etc.

Structs

BallJoint
A joint that removes all relative linear motion between a pair of points on two bodies.

CCDSolver
Solver responsible for performing motion-clamping on fast-moving bodies.

FixedJoint
A joint that prevents all relative movement between two bodies.

IntegrationParameters
Parameters for a time-step of the physics engine.

IslandManager
Structure responsible for maintaining the set of active rigid-bodies, and putting non-moving rigid-bodies to sleep to save computation times.

Joint
A joint attached to two bodies.

JointHandle
The unique identifier of a joint added to the joint set. The unique identifier of a collider added to a collider set.

JointSet
A set of joints that can be handled by a physics World.

MassProperties
The local mass properties of a rigid-body.

PrismaticJoint
A joint that removes all relative motion between two bodies, except for the translations along one axis.

RigidBodyActivation
The rb_activation status of a body.

RigidBodyCcd
Information used for Continuous-Collision-Detection.

RigidBodyChanges
Flags describing how the rigid-body has been modified by the user.

RigidBodyColliders
The set of colliders attached to this rigid-bodies.

RigidBodyDamping
Damping factors to progressively slow down a rigid-body.

RigidBodyDominance
The dominance groups of a rigid-body.

RigidBodyForces
The user-defined external forces applied to this rigid-body.

RigidBodyHandle
The unique handle of a rigid body added to a RigidBodySet.

RigidBodyIds
Internal identifiers used by the physics engine.

RigidBodyMassProps
The mass properties of this rigid-bodies.

RigidBodyMassPropsFlags
Flags affecting the behavior of the constraints solver for a given contact manifold.

RigidBodyPosition
The position of this rigid-body.

RigidBodyVelocity
The velocities of this rigid-body.

Enums

CoefficientCombineRule
Rules used to combine two coefficients.

JointParams
An enum grouping all possible types of joints.

RigidBodyType
The status of a body, governing the way it is affected by external forces.

SpringModel
The spring-like model used for constraints resolution.

Type Definitions

BodyStatus	Deprecated
The type of a body, governing the way it is affected by external forces.