rapier2d
Module geometry

Structures related to geometry: colliders, shapes, etc.

Structs

ActiveCollisionTypes

Flags affecting whether or not collision-detection happens between two colliders depending on the type of rigid-bodies they are attached to.

BroadPhaseMultiSap

A broad-phase combining a Hierarchical Grid and Sweep-and-Prune.

Collider

A geometric entity that can be attached to a body so it can be affected by contacts and proximity queries.

ColliderBroadPhaseData

Data associated to a collider that takes part to a broad-phase algorithm.

ColliderBuilder

A structure responsible for building a new collider.

ColliderChanges

Flags describing how the collider has been modified by the user.

ColliderFlags

A set of flags for controlling collision/intersection filtering, modification, and events.

ColliderHandle

The unique identifier of a collider added to a collider set.

ColliderMaterial

The constraints solver-related properties of this collider (friction, restitution, etc.)

ColliderPair

A pair of collider handles.

ColliderParent

Information about the rigid-body this collider is attached to.

ColliderPosition

The position of a collider.

ColliderSet

A set of colliders that can be handled by a physics World.

CollisionEventFlags

Flags providing more information regarding a collision event.

Compound

A compound shape with an aabb bounding volume.

ContactData

A single contact between two collider.

ContactForceEvent

Event occurring when the sum of the magnitudes of the contact forces between two colliders exceed a threshold.

ContactManifoldData

A contact manifold between two colliders.

ContactPair

The description of all the contacts between a pair of colliders.

ConvexPolygon

A 2D convex polygon.

Group

A bit mask identifying groups for interaction.

HalfSpace

A half-space delimited by an infinite plane.

InteractionGraph

A graph where nodes are collision objects and edges are contact or proximity algorithms.

InteractionGroups

Pairwise filtering using bit masks.

IntersectionPair

The description of all the contacts between a pair of colliders.

NarrowPhase

The narrow-phase responsible for computing precise contact information between colliders.

PackedFeatureId

A feature id where the feature type is packed into the same value as the feature index.

PolygonalFeature

A polygonal feature representing the local polygonal approximation of a vertex, or face, of a convex shape.

Polyline

A polyline.

RoundShape

A shape with rounded borders.

SharedShape

The shape of a collider.

SolverContact

A contact seen by the constraints solver for computing forces.

SolverFlags

Flags affecting the behavior of the constraints solver for a given contact manifold.

TopoFace

A face of a triangle-mesh’s half-edge topology.

TopoHalfEdge

A half-edge of a triangle-mesh’s half-edge topology.

TopoVertex

A vertex of a triangle-mesh’s half-edge topology.

TrackedContact

A single contact between two shape.

TriMesh

A triangle mesh.

TriMeshConnectedComponents

The connected-components of a triangle mesh.

TriMeshFlags

The status of the cell of an heightfield.

TriMeshTopology

The half-edge topology information of a triangle mesh.

TrianglePseudoNormals

The pseudo-normals of a triangle providing approximations of its feature’s normal cones.

Enums

BroadPhasePairEvent

An event emitted by the broad-phase.

ColliderEnabled

Enum indicating whether or not a collider is enabled.

ColliderMassProps

The mass-properties of a collider.

ColliderType

The type of collider.

CollisionEvent

Events occurring when two colliders start or stop colliding

FeatureId

An identifier of a feature of a convex polyhedron.

MeshConverter

Determines how meshes (generally when loaded from a file) are converted into Rapier colliders.

MeshConverterError

Error that can be generated by the MeshConverter.

SegmentPointLocation

Logical description of the location of a point on a triangle.

ShapeType

Enum representing the type of a shape.

TopologyError

Indicated an inconsistency in the topology of a triangle mesh.

TriangleOrientation

Orientation of a triangle.

TrianglePointLocation

Description of the location of a point on a triangle.

TypedShape

Enum representing the shape with its actual type

Traits

BoundingVolume

Trait of bounding volumes.

BroadPhase

Trait implemented by broad-phase algorithms supported by Rapier.

PointQuery

Trait of objects that can be tested for point inclusion and projection.

PointQueryWithLocation

Returns shape-specific info in addition to generic projection information

PolygonalFeatureMap

Trait implemented by convex shapes with features with polyhedral approximations.

RayCast

Traits of objects which can be transformed and tested for intersection with a ray.

Shape

Trait implemented by shapes usable by Rapier.

SimdCompositeShape

Trait implemented by shapes composed of multiple simpler shapes.

SupportMap

Traits of convex shapes representable by a support mapping function.

TypedSimdCompositeShape

Type Aliases

Aabb

An axis-aligned bounding box.

Ball

A ball shape.

Capsule

A capsule shape.

ColliderGraphIndex

Index of a node of the interaction graph.

ColliderShape

The shape of a collider.

Contact

A contact between two colliders.

ContactManifold

A contact manifold between two colliders.

Cuboid

A cuboid shape.

DefaultBroadPhase

The default broad-phase implementation recommended for general-purpose usage.

HeightField

A heightfield shape.

HeightFieldCellStatus

Indicates if a cell of a heightfield is removed or not. Set this to false for a removed cell.

PointProjection

The projection of a point on a collider.

Ray

A ray that can be cast against colliders.

RayIntersection

The intersection between a ray and a collider.

RigidBodyGraphIndex

Index of a node of the interaction graph.

RoundConvexPolygon

A convex polygon dilated by a sphere (so it has round corners).

RoundCuboid

A cuboid dilated by a sphere (so it has round corners).

RoundTriangle

A triangle dilated by a sphere (so it has round corners).

Segment

A segment shape.

ShapeCastHit

The result of a shape-cast between two shapes.

TemporaryInteractionIndex

Temporary index to and edge of the interaction graph.

Triangle

A triangle shape.