rapier2d::geometry
Struct SharedShape

pub struct SharedShape(pub Arc<dyn Shape>);

The shape of a collider.

Tuple Fields
0: Arc<dyn Shape>
Implementations

impl SharedShape

pub fn new(shape: impl Shape) -> SharedShape

Wraps the given shape as a shared shape.

pub fn make_mut(&mut self) -> &mut (dyn Shape + 'static)

If this shape is shared, then the content of self is cloned into a unique instance, and a mutable reference to that instance is returned.

pub fn compound( shapes: Vec<(Isometry<f32, Unit<Complex<f32>>, 2>, SharedShape)>, ) -> SharedShape

Initialize a compound shape defined by its subshapes.

pub fn ball(radius: f32) -> SharedShape

Initialize a ball shape defined by its radius.

pub fn halfspace( outward_normal: Unit<Matrix<f32, Const<2>, Const<1>, ArrayStorage<f32, 2, 1>>>, ) -> SharedShape

Initialize a plane shape defined by its outward normal.

pub fn cuboid(hx: f32, hy: f32) -> SharedShape

Initialize a cuboid shape defined by its half-extents.

pub fn round_cuboid(hx: f32, hy: f32, border_radius: f32) -> SharedShape

Initialize a round cuboid shape defined by its half-extents and border radius.

pub fn capsule( a: OPoint<f32, Const<2>>, b: OPoint<f32, Const<2>>, radius: f32, ) -> SharedShape

Initialize a capsule shape from its endpoints and radius.

pub fn capsule_x(half_height: f32, radius: f32) -> SharedShape

Initialize a capsule shape aligned with the x axis.

pub fn capsule_y(half_height: f32, radius: f32) -> SharedShape

Initialize a capsule shape aligned with the y axis.

pub fn segment( a: OPoint<f32, Const<2>>, b: OPoint<f32, Const<2>>, ) -> SharedShape

Initialize a segment shape from its endpoints.

pub fn triangle( a: OPoint<f32, Const<2>>, b: OPoint<f32, Const<2>>, c: OPoint<f32, Const<2>>, ) -> SharedShape

Initializes a triangle shape.

pub fn round_triangle( a: OPoint<f32, Const<2>>, b: OPoint<f32, Const<2>>, c: OPoint<f32, Const<2>>, border_radius: f32, ) -> SharedShape

Initializes a triangle shape with round corners.

pub fn polyline( vertices: Vec<OPoint<f32, Const<2>>>, indices: Option<Vec<[u32; 2]>>, ) -> SharedShape

Initializes a polyline shape defined by its vertex and index buffers.

If no index buffer is provided, the polyline is assumed to describe a line strip.

pub fn trimesh( vertices: Vec<OPoint<f32, Const<2>>>, indices: Vec<[u32; 3]>, ) -> SharedShape

Initializes a triangle mesh shape defined by its vertex and index buffers.

pub fn trimesh_with_flags( vertices: Vec<OPoint<f32, Const<2>>>, indices: Vec<[u32; 3]>, flags: TriMeshFlags, ) -> SharedShape

Initializes a triangle mesh shape defined by its vertex and index buffers and pre-processing flags.

pub fn convex_decomposition( vertices: &[OPoint<f32, Const<2>>], indices: &[[u32; 2]], ) -> SharedShape

Initializes a compound shape obtained from the decomposition of the given trimesh (in 3D) or polyline (in 2D) into convex parts.

pub fn round_convex_decomposition( vertices: &[OPoint<f32, Const<2>>], indices: &[[u32; 2]], border_radius: f32, ) -> SharedShape

Initializes a compound shape obtained from the decomposition of the given trimesh (in 3D) or polyline (in 2D) into convex parts dilated with round corners.

pub fn convex_decomposition_with_params( vertices: &[OPoint<f32, Const<2>>], indices: &[[u32; 2]], params: &VHACDParameters, ) -> SharedShape

Initializes a compound shape obtained from the decomposition of the given trimesh (in 3D) or polyline (in 2D) into convex parts.

pub fn round_convex_decomposition_with_params( vertices: &[OPoint<f32, Const<2>>], indices: &[[u32; 2]], params: &VHACDParameters, border_radius: f32, ) -> SharedShape

Initializes a compound shape obtained from the decomposition of the given trimesh (in 3D) or polyline (in 2D) into convex parts dilated with round corners.

pub fn convex_hull(points: &[OPoint<f32, Const<2>>]) -> Option<SharedShape>

Creates a new shared shape that is the convex-hull of the given points.

pub fn convex_polyline( points: Vec<OPoint<f32, Const<2>>>, ) -> Option<SharedShape>

Creates a new shared shape that is a convex polygon formed by the given set of points assumed to form a convex polyline (no convex-hull will be automatically computed).

pub fn round_convex_hull( points: &[OPoint<f32, Const<2>>], border_radius: f32, ) -> Option<SharedShape>

Creates a new shared shape with rounded corners that is the convex-hull of the given points, dilated by border_radius.

pub fn round_convex_polyline( points: Vec<OPoint<f32, Const<2>>>, border_radius: f32, ) -> Option<SharedShape>

Creates a new shared shape with round corners that is a convex polygon formed by the given set of points assumed to form a convex polyline (no convex-hull will be automatically computed).

pub fn heightfield( heights: Matrix<f32, Dyn, Const<1>, VecStorage<f32, Dyn, Const<1>>>, scale: Matrix<f32, Const<2>, Const<1>, ArrayStorage<f32, 2, 1>>, ) -> SharedShape

Initializes a heightfield shape defined by its set of height and a scale factor along each coordinate axis.