Trait rapier2d::geometry::SimdCompositeShape

pub trait SimdCompositeShape {
    pub fn map_part_at(
        &self, 
        shape_id: u32, 
        f: &mut dyn FnMut(Option<&Isometry<f32, Unit<Complex<f32>>, 2_usize>>, &(dyn Shape + 'static))
    );

    pub fn quadtree(&self) -> &QBVH<u32>;
}

Trait implemented by shapes composed of multiple simpler shapes.

A composite shape is composed of several shapes. For example, this can be a convex decomposition of a concave shape; or a triangle-mesh.

Required methods

pub fn map_part_at(
&self,
shape_id: u32,
f: &mut dyn FnMut(Option<&Isometry<f32, Unit<Complex<f32>>, 2_usize>>, &(dyn Shape + 'static))
)

Applies a function to one sub-shape of this composite shape.

pub fn quadtree(&self) -> &QBVH<u32>

Gets the acceleration structure of the composite shape.

Loading content...