rapier2d
Module pipeline

Structure for combining the various physics components to perform an actual simulation.

Modules

query_pipeline_generators

Structs implementing QbvhDataGenerator<ColliderHandle> to be used with QueryPipeline::update_with_generator.

Structs

ActiveEvents

Flags affecting the events generated for this collider.

ActiveHooks

Flags affecting the behavior of the constraints solver for a given contact manifold.

ChannelEventCollector

A collision event handler that collects events into a crossbeam channel.

CollisionPipeline

The collision pipeline, responsible for performing collision detection between colliders.

ContactModificationContext

Context given to custom contact modifiers to modify the contacts seen by the constraints solver.

DebugRenderMode

Flags indicating what part of the physics engine should be rendered by the debug-renderer.

DebugRenderPipeline

Pipeline responsible for rendering the state of the physics engine for debugging purpose.

DebugRenderStyle

Style used for computing colors when rendering the scene.

PairFilterContext

Context given to custom collision filters to filter-out collisions.

PhysicsPipeline

The physics pipeline, responsible for stepping the whole physics simulation.

QueryFilter

A filter that describes what collider should be included or excluded from a scene query.

QueryFilterFlags

Flags for excluding whole sets of colliders from a scene query.

QueryPipeline

A pipeline for performing queries on all the colliders of a scene.

Enums

DebugRenderObject

The object currently being rendered by the debug-renderer.

Traits

DebugRenderBackend

Trait implemented by graphics backends responsible for rendering the physics scene.

EventHandler

Trait implemented by structures responsible for handling events generated by the physics engine.

PhysicsHooks

User-defined functions called by the physics engines during one timestep in order to customize its behavior.

Type Aliases

DebugColor

A color for debug-rendering.